requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.8.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coverage==6.4.2
cryptography==38.0.1
Django==4.0.6
django-js-asset==2.0.0
django-mptt==0.13.4
factory-boy==3.2.1
Faker==14.2.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.1
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
python-decouple==3.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
stripe==4.1.0
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
